requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7
coverage==7.8.0
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-utils==0.8.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
fsspec==2025.3.2
gevent==25.4.1
geventhttpclient==2.3.3
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
locust==2.35.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
mypy-extensions==1.0.0
narwhals==1.35.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
pluggy==1.5.0
protobuf==5.29.4
psutil==5.9.8
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
torch==2.6.0
tornado==6.4.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
watchdog==6.0.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
